requirements.txt
mesa>=3.5.0
cupy-cuda12x>=14.0.0
numba>=0.64.0
numpy>=2.0
scipy>=1.14
pandas>=2.2
geopandas>=1.1
matplotlib>=3.9
seaborn>=0.13
networkx>=3.3
shapely>=2.1
jupyter>=1.1
